split scanner and modal component
